Capitalization function, loop for all test cases
Added a function to capitalize the first letter of the input sentence. The noun replace function only adds "you". Also, created a loop to check all test cases at once. It makes a list of lists for test cases with more than one possible outcome. For ease of use, we may want to just make each of them an individual list entry.