Fixed pronoun and verb replace functions
The pronoun replace function was only replacing the last pronoun in the sentence, and it turns out the verb function was doing the same. I added the recursion back into the verb function and updated the pronoun function to have recursion as well. Also broke out the pronoun type counter into its own function.