Removed fluff from API call loop
The API call loop had a while loop in it and a counter variable j, which was unnecessary, and I think somehow it was actually causing problems. It's gone now, and the API calls seem to be happening 1 for 1, which is good because before they were randomly 1 for 1 or 5 for 1 or some other random number. Too unpredictable for my limited number of API calls.