Reorganized repo structure
Split the repo into two main pieces: the part that builds the dataset and the web app. The dataset portion is further subdivided into data retrieval and data transformation. The result of the data transformation will be put into the data folder in the web app side of the repo for use at runtime. This is the whole purpose of this branch: to complete all language transformations ahead of time and provide processed data to the app so it can be presented as quickly as possible to the user.