Changed data to ID keys instead of titles
Some movies have more than one title for a single ID. Originally, I was using titles as the dictionary key, but that caused problems with the API that would scale as the dataset grew. The ID is less interpretable at a glance, but it solves the duplicate issue and makes the API calls work correctly.