Export transformed data for annotation
Transformed plots are now exported to a CSV file. Each plot is its own line along with the movie id and title. This CSV should be imported into Tortus for quality annotation.

Also, added a progress counter to the transformation loop and removed some package imports that weren't being used.